fixed donation route
