+++ multer.js
@@ -34,7 +34,7 @@ function checkFileType(file, cb) {
   if (mimetype && extname) {
     return cb(null, true);
   } else {
-    cb('Error: Images Only!');
+    cb(null, false);
   }
 }
