+++ views/partials/donation_form.handlebars
@@ -1,9 +1,4 @@
 
-{{#if msg}}
-  {{'hey'}}
-{{else}}
-  {{msg}}
-{{/if}}
 {{{msg}}}
 
 <form id="donation_form" enctype="multipart/form-data" action={{action}} method={{method}}>
